ISSUE #883: Rename RewindItemsPerYear to RewindItemsPerGroup
