+++ tests/Domain/Strava/Rewind/FindPersonalRecordsPerMonth/FindPersonalRecordsPerMonthQueryHandlerTest.php
@@ -5,7 +5,6 @@ namespace App\Tests\Domain\Strava\Rewind\FindPersonalRecordsPerMonth;
 use App\Domain\Strava\Activity\ActivityId;
 use App\Domain\Strava\Activity\ActivityWithRawData;
 use App\Domain\Strava\Activity\ActivityWithRawDataRepository;
-use App\Domain\Strava\Calendar\Month;
 use App\Domain\Strava\Gear\GearId;
 use App\Domain\Strava\Rewind\FindPersonalRecordsPerMonth\FindPersonalRecordsPerMonth;
 use App\Domain\Strava\Rewind\FindPersonalRecordsPerMonth\FindPersonalRecordsPerMonthQueryHandler;
@@ -80,8 +79,8 @@ class FindPersonalRecordsPerMonthQueryHandlerTest extends ContainerTestCase
         $response = $this->queryHandler->handle(new FindPersonalRecordsPerMonth(Years::fromArray([Year::fromInt(2024)])));
         $this->assertEquals(
             [
-                [Month::fromDate(SerializableDateTime::fromString('2024-03-01')), 3],
-                [Month::fromDate(SerializableDateTime::fromString('2024-01-01')), 5],
+                [3, 3],
+                [1, 5],
             ],
             $response->getPersonalRecordsPerMonth(),
         );
